pyproject.toml
[project]
name = "techxchange_hackathon_neuralnexis"
version = "0.1.0"
description = "TechXchange_Hackathon_NeuralNexis using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.159.0,<1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.1.1",
    "praw>=7.8.1",
    "duckduckgo-search>=8.1.1",
    "wikipedia>=1.4.0",
]

[project.scripts]
techxchange_hackathon_neuralnexis = "techxchange_hackathon_neuralnexis.main:run"
run_crew = "techxchange_hackathon_neuralnexis.main:run"
train = "techxchange_hackathon_neuralnexis.main:train"
replay = "techxchange_hackathon_neuralnexis.main:replay"
test = "techxchange_hackathon_neuralnexis.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
